Let brl_buf_state.brl_check() identify mixed state
When some braille pattern in the JSON file is prefix of another braille
pattern, a braille input can be the intermediate and the final state at the
same time. The method now returns a pair (string, flag). A nonempty string
indicates the final output, while the flag is True for an intermediate state.