Implement braille input via numpad keys
This feature enables usage with fingers that cannot press several keys at
the same time.

Keys 0 to 8 represent space and dots 1 to 8 respectively. 9 is used to
cancel dots (and/or space) previously recorded by numpad keys. The decimal
point (.) is used to commit these dots (and/or space).

Any press of non-ignored keys on the main keyboard interrupts the record
process of numpad keys.