Replace beep_typo.wav with its original version
To satisfy users' practice, beep_typo.wav is the same as beep.wav of G-Mouse
(視窗導盲鼠系統). However, the format of beep.wav is not compatible
with Python, so beep_typo.wav is a version of beep.wav with different
format. Now, because PlaySound is a native function directly offered by
WinAPI, this file is exactly identical to beep.wav of G-Mouse.