runtime_state.py: Add foreground_process_change_notify method
It only works when ONE_CBRLKB_TOGGLE_STATE is turned off. A nonzero return
value indicates enable (1) or disable (-1).

Name of the method is not good enough, because it is actually called
when
the foreground window changes. If a process owns more than one window,
e.g. web browser, the method is called even when the user switches the
focus between the two windows.

[!] Partial implementation of ONE_CBRLKB_TOGGLE_STATE option.