Take "noBraille" as the source of dummy braille inputs
The addon attempts to bring least influence to other parts of NVDA. The
behavior of the physical braille keyboard on a braille display should be
defined by its driver completely. Thus, source of dummy braille inputs is
altered to name of the only dummy braille display driver so that newly
inserted braille-key shortcuts are independent of old ones in the driver.

However, NVDA versions less than 2018.3 force gestures from different
braille display to show name of the current working braille display as
their source. Behavior of the addon running on old versions of NVDA is
therefore different from that running on new versions of NVDA.