Implement script_copyBRLdisplayCells
Key shortcut: [NVDA]+[PrintScreen]

There are two modes, Raw Text and Braille Unicode.

1. Raw Text
   Press once: Copy the raw text of all visible braille regions, including
      attribute information such as the control type and the key shortcut.
   Press twice: Copy the raw text of all visible and hidden braille
      regions.
2. Unicode Braille
   Press once: Copy the content of the current window of the braille
      display in braille Unicode (U+2800-U+28FF).
   Press twice: Copy the content of all braille regions in Unicode
      braille.
3. Press three times: Switch between the two modes.

This may be an alternative for users who would like to copy the raw text
of all UI messages. However, the addon is not responsible for copying the
content of NVDA speech. Besides, this feature will help a lot to tutorial
material editing.

There is one additional space character at the end of the output string
of a braille region with a cursor. Though the addon tries to detect and
remove it, I think that NVDA may change this behavior in the future.

The original feature proposal was from 嘯傲俠羽 <crazy@mail.batol.net>.