__init__.py: Rewrite the input mode inference rules
It is a subset of the input mode inference rule.

If the current keyboard layout is recognized and its corresponding input
processor CLSID is not the same as that of the default language profile,
then guess the native input mode. Note that the rule becomes independent
of the Windows version.

If the CLSID of the default profile is GUID_null and the layout string is
not determined yet, then guess the alphanumeric mode.

With a known layout string, the guessed result simply depends on the CLSID
of the default language profile (is or is not GUID_null).

If none of the above rules fits, then guess the alphanumeric input mode.