keyboard.py: Improve hack_compositionUpdate() again
[!] Assume no alphanumeric character should be typed on the native mode.

On composition string change, take action according to the following 3
cases of the inserted character(s).

1. All bopomofo symbols:

It occurs when (1) composition of a new Han character, or (2) composition
of a new bopomofo character by the hexadecimal representation.

Case (1). Call the original compositionUpdate() with announce=False.
Case (2). Call the original compositionUpdate() as expected.

2. All alphanumeric characters:

It occurs during the case (2) above. The original compositionUpdate()
will not be called except the case that only a '`' is inserted into
the composition string (with announce=False). The exception exists to
activate the case (2) above.

3. Otherwise, the original compositionUpdate() is called as expected.