__init__.py: Add script_internalCodeBRL()
The default gesture is 137 + space.

When entering the internal code braille state, the content on the braille
display becomes the internal code braille based on UTF-8. The same gesture
is used to cancel the state.

The braille pattern is built by NABCCX. The text is encoded by UTF-8, and
its each byte is translated into braille via NABCCX rules. Therefore, each
character has its unique braille pattern, but it is not rememberable at all.

Note: The behavior of Python 2 is different from that of Python 3.