Use brailleInput.handler.sendChars() to send Unicode characters
It sends the character via KEYBDINPUT with KEYEVENTF_UNICODE flag. The
message can be processed by the input composition object to generate the
specified character on it. It provides an approach without `uhhhh string.

In the IME JSON file, ALLOW_NON_BMP_SYMBOLS is redefined. If the value is
false, then the addon rejects to send any non-BMP character when the input
composition object is focused.

What the method sends are not emulated keyboard messages, so do not send
ASCII characters via it.