Rewrite the computer keyboard emulation
The new keyboard_hook.py contains the following moved from __init__.py:

- DummyBrailleInputGesture class.
- The keyboard callbacks and all related variables.

The new rules of the key-down and key-up callbacks:

All of the key messages are classified into 3 categories in the key-down
callback: passed, trapped, and pending.

Passed keys: A key is passed on satisfaction of any condition below.

1. NVDA is on the browse mode.
2. The focus is some special edit control in the settings panel.
3. Shift with previously pressed modifier.
4. The other modifier pressed when Shift is not down.
5. The key with an invalid character code returned by MapVirtualKeyExW().
6. The key modified by any modifier key.
7. The modified or undefined NumPad keys.
8. The other key which is not accepted.

Trapped keys: A key is trapped on satisfaction of any condition below.

1. Shift without previously pressed modifier.
2. The other modifier pressed when Shift is down.
3. The key modified by some trapped modifier.
4. The processed NumPad key.

When self._ignored_keys is not None, all of the key-down messages are not
processed. The state, called a key-pass session, ends after release of
the keys recorded in self._ignored_keys.

The other cases result in pending keys.

The major difference of the new version is to classify the keys by the
action taken by the key-up callback.